Edit README.md file, babel config, move all modules in src/modules, also add description in package.json and edit index.js into src/index.js
