+++ .gitignore
@@ -3,4 +3,4 @@ dist/
 *.vscode
 *.git
 *.log
-.DS-Store
+.DS_Store
