+++ README.md
@@ -4,3 +4,11 @@
 ### https://github.com/7StaSH7/Sprint-11
 ## Ссылка на сам проект:
 ### https://7stash7.github.io/Sprint-11/ 
+## Установка:
+* [Установите Node.js](https://nodejs.org/en/download/) 
+* Скопируйте репозиторий к себе на компьютер
+* Зайдите в проект и напишите `npm install`
+## Запуск проекта:
+* `npm run dev` - запуск в режиме разработчика
+* `npm run build` - запуск финальной сборки
+* `npm run deploy` - размещение проекта на GitHub Pages
