+++ package.json
@@ -2,7 +2,7 @@
   "name": "sprint-11",
   "version": "1.0.0",
   "description": "Собранный код 9 проектной работы",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
     "build": "cross-env NODE_ENV=production rimraf dist && webpack --mode production",
     "dev": "cross-env NODE_ENV=development webpack-dev-server --mode development --open --watch",
